allow only known Kodi versions
before it has done a fallback to Leia if it was not known or not set.
This can then if Kodi N* is coming a bit difficult as it think it is for
Leia.

If it then works for build and a outside create a release, becomes then a
Leia replaced by a N* version of addon.

As long there are only 2 versions inside here it is easy, then more like
a yes/no, but after them with 3 or more a bit harder.